+++ druid/src/core.rs
@@ -44,7 +44,7 @@ pub(crate) type CommandQueue = VecDeque<(Target, Command)>;
 ///
 /// [`update`]: trait.Widget.html#tymethod.update
 pub struct WidgetPod<T, W> {
-    state: BaseState,
+    state: WidgetState,
     old_data: Option<T>,
     env: Option<Env>,
     inner: W,
@@ -68,7 +68,7 @@ pub struct WidgetPod<T, W> {
 /// [`paint`]: trait.Widget.html#tymethod.paint
 /// [`WidgetPod`]: struct.WidgetPod.html
 #[derive(Clone)]
-pub(crate) struct BaseState {
+pub(crate) struct WidgetState {
     pub(crate) id: WidgetId,
     /// The frame of this widget in its parents coordinate space.
     /// This should always be set; it is only an `Option` so that we
@@ -138,7 +138,7 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
     /// so it can participate in layout and event flow. The process of
     /// adding a child widget to a container should call this method.
     pub fn new(inner: W) -> WidgetPod<T, W> {
-        let mut state = BaseState::new(inner.id().unwrap_or_else(WidgetId::next));
+        let mut state = WidgetState::new(inner.id().unwrap_or_else(WidgetId::next));
         state.children_changed = true;
         state.needs_layout = true;
         WidgetPod {
@@ -152,7 +152,7 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
 
     /// Read-only access to state. We don't mark the field as `pub` because
     /// we want to control mutation.
-    pub(crate) fn state(&self) -> &BaseState {
+    pub(crate) fn state(&self) -> &WidgetState {
         &self.state
     }
 
@@ -212,7 +212,7 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
         if old_size.is_none() || old_size.unwrap() != new_size {
             let mut child_ctx = LifeCycleCtx {
                 command_queue: ctx.command_queue,
-                base_state: &mut self.state,
+                widget_state: &mut self.state,
                 window_id: ctx.window_id,
                 window: ctx.window,
             };
@@ -236,7 +236,7 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
         }
 
         if needs_merge {
-            ctx.base_state.merge_up(&mut self.state);
+            ctx.widget_state.merge_up(&mut self.state);
         }
     }
 
@@ -335,7 +335,7 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
     fn set_hot_state(
         child: &mut W,
         command_queue: &mut CommandQueue,
-        child_state: &mut BaseState,
+        child_state: &mut WidgetState,
         window_id: WindowId,
         window: &WindowHandle,
         rect: Rect,
@@ -352,7 +352,7 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
             let hot_changed_event = LifeCycle::HotChanged(child_state.is_hot);
             let mut child_ctx = LifeCycleCtx {
                 command_queue,
-                base_state: child_state,
+                widget_state: child_state,
                 window_id,
                 window,
             };
@@ -390,7 +390,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
             window_id: ctx.window_id,
             z_ops: Vec::new(),
             region: ctx.region.clone(),
-            base_state: &self.state,
+            widget_state: &self.state,
             focus_widget: ctx.focus_widget,
             depth: ctx.depth,
         };
@@ -518,7 +518,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
         };
         let mut child_ctx = LayoutCtx {
             command_queue: ctx.command_queue,
-            base_state: &mut self.state,
+            widget_state: &mut self.state,
             window_id: ctx.window_id,
             window: ctx.window,
             text_factory: ctx.text_factory,
@@ -526,7 +526,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
         };
         let size = self.inner.layout(&mut child_ctx, bc, data, env);
 
-        ctx.base_state.merge_up(&mut child_ctx.base_state);
+        ctx.widget_state.merge_up(&mut child_ctx.widget_state);
 
         if size.width.is_infinite() {
             let name = self.widget().type_name();
@@ -587,14 +587,14 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
             command_queue: ctx.command_queue,
             window: ctx.window,
             window_id: ctx.window_id,
-            base_state: &mut self.state,
+            widget_state: &mut self.state,
             is_handled: false,
             is_root: false,
             focus_widget: ctx.focus_widget,
             app_data_type: ctx.app_data_type,
         };
 
-        let rect = child_ctx.base_state.layout_rect.unwrap_or_default();
+        let rect = child_ctx.widget_state.layout_rect.unwrap_or_default();
 
         // Note: could also represent this as `Option<Event>`.
         let mut recurse = true;
@@ -604,7 +604,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                     let hot_changed = WidgetPod::set_hot_state(
                         &mut self.inner,
                         child_ctx.command_queue,
-                        child_ctx.base_state,
+                        child_ctx.widget_state,
                         child_ctx.window_id,
                         child_ctx.window,
                         rect,
@@ -624,7 +624,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                         Target::Widget(id) => {
                             // Recurse when the target widget could be our descendant.
                             // The bloom filter we're checking can return false positives.
-                            recurse = child_ctx.base_state.children.may_contain(id);
+                            recurse = child_ctx.widget_state.children.may_contain(id);
                             Event::Internal(InternalEvent::TargetedCommand(*target, cmd.clone()))
                         }
                         Target::Global => {
@@ -634,8 +634,8 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 }
                 InternalEvent::RouteTimer(token, widget_id) => {
                     let widget_id = *widget_id;
-                    if widget_id != child_ctx.base_state.id {
-                        recurse = child_ctx.base_state.children.may_contain(&widget_id);
+                    if widget_id != child_ctx.widget_state.id {
+                        recurse = child_ctx.widget_state.children.may_contain(&widget_id);
                         Event::Internal(InternalEvent::RouteTimer(*token, widget_id))
                     } else {
                         Event::Timer(*token)
@@ -652,7 +652,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 WidgetPod::set_hot_state(
                     &mut self.inner,
                     child_ctx.command_queue,
-                    child_ctx.base_state,
+                    child_ctx.widget_state,
                     child_ctx.window_id,
                     child_ctx.window,
                     rect,
@@ -660,7 +660,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                     data,
                     env,
                 );
-                recurse = had_active || child_ctx.base_state.is_hot;
+                recurse = had_active || child_ctx.widget_state.is_hot;
                 let mut mouse_event = mouse_event.clone();
                 mouse_event.pos -= rect.origin().to_vec2();
                 Event::MouseDown(mouse_event)
@@ -669,7 +669,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 WidgetPod::set_hot_state(
                     &mut self.inner,
                     child_ctx.command_queue,
-                    child_ctx.base_state,
+                    child_ctx.widget_state,
                     child_ctx.window_id,
                     child_ctx.window,
                     rect,
@@ -677,7 +677,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                     data,
                     env,
                 );
-                recurse = had_active || child_ctx.base_state.is_hot;
+                recurse = had_active || child_ctx.widget_state.is_hot;
                 let mut mouse_event = mouse_event.clone();
                 mouse_event.pos -= rect.origin().to_vec2();
                 Event::MouseUp(mouse_event)
@@ -686,7 +686,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 let hot_changed = WidgetPod::set_hot_state(
                     &mut self.inner,
                     child_ctx.command_queue,
-                    child_ctx.base_state,
+                    child_ctx.widget_state,
                     child_ctx.window_id,
                     child_ctx.window,
                     rect,
@@ -697,7 +697,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 // MouseMove is recursed even if the widget is not active and not hot,
                 // but was hot previously. This is to allow the widget to respond to the movement,
                 // e.g. drag functionality where the widget wants to follow the mouse.
-                recurse = had_active || child_ctx.base_state.is_hot || hot_changed;
+                recurse = had_active || child_ctx.widget_state.is_hot || hot_changed;
                 let mut mouse_event = mouse_event.clone();
                 mouse_event.pos -= rect.origin().to_vec2();
                 Event::MouseMove(mouse_event)
@@ -706,7 +706,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 WidgetPod::set_hot_state(
                     &mut self.inner,
                     child_ctx.command_queue,
-                    child_ctx.base_state,
+                    child_ctx.widget_state,
                     child_ctx.window_id,
                     child_ctx.window,
                     rect,
@@ -714,7 +714,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                     data,
                     env,
                 );
-                recurse = had_active || child_ctx.base_state.is_hot;
+                recurse = had_active || child_ctx.widget_state.is_hot;
                 let mut mouse_event = mouse_event.clone();
                 mouse_event.pos -= rect.origin().to_vec2();
                 Event::Wheel(mouse_event)
@@ -732,7 +732,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 Event::Paste(e.clone())
             }
             Event::Zoom(zoom) => {
-                recurse = had_active || child_ctx.base_state.is_hot;
+                recurse = had_active || child_ctx.widget_state.is_hot;
                 Event::Zoom(*zoom)
             }
             Event::Timer(token) => {
@@ -742,12 +742,12 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
             Event::Command(cmd) => Event::Command(cmd.clone()),
         };
         if recurse {
-            child_ctx.base_state.has_active = false;
+            child_ctx.widget_state.has_active = false;
             self.inner.event(&mut child_ctx, &child_event, data, env);
-            child_ctx.base_state.has_active |= child_ctx.base_state.is_active;
+            child_ctx.widget_state.has_active |= child_ctx.widget_state.is_active;
         };
 
-        ctx.base_state.merge_up(&mut child_ctx.base_state);
+        ctx.widget_state.merge_up(&mut child_ctx.widget_state);
         ctx.is_handled |= child_ctx.is_handled;
     }
 
@@ -844,7 +844,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
 
         let mut child_ctx = LifeCycleCtx {
             command_queue: ctx.command_queue,
-            base_state: &mut self.state,
+            widget_state: &mut self.state,
             window_id: ctx.window_id,
             window: ctx.window,
         };
@@ -857,14 +857,14 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
             self.inner.lifecycle(&mut child_ctx, event, data, env);
         }
 
-        ctx.base_state.merge_up(&mut self.state);
+        ctx.widget_state.merge_up(&mut self.state);
 
         // we need to (re)register children in case of one of the following events
         match event {
             LifeCycle::WidgetAdded | LifeCycle::Internal(InternalLifeCycle::RouteWidgetAdded) => {
                 self.state.children_changed = false;
-                ctx.base_state.children = ctx.base_state.children.union(self.state.children);
-                ctx.base_state.focus_chain.extend(&self.state.focus_chain);
+                ctx.widget_state.children = ctx.widget_state.children.union(self.state.children);
+                ctx.widget_state.focus_chain.extend(&self.state.focus_chain);
                 ctx.register_child(self.id());
             }
             _ => (),
@@ -891,7 +891,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
 
         let mut child_ctx = UpdateCtx {
             window: ctx.window,
-            base_state: &mut self.state,
+            widget_state: &mut self.state,
             window_id: ctx.window_id,
             command_queue: ctx.command_queue,
         };
@@ -901,7 +901,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
         self.old_data = Some(data.clone());
         self.env = Some(env.clone());
 
-        ctx.base_state.merge_up(&mut self.state)
+        ctx.widget_state.merge_up(&mut self.state)
     }
 }
 
@@ -915,9 +915,9 @@ impl<T, W: Widget<T> + 'static> WidgetPod<T, W> {
     }
 }
 
-impl BaseState {
-    pub(crate) fn new(id: WidgetId) -> BaseState {
-        BaseState {
+impl WidgetState {
+    pub(crate) fn new(id: WidgetId) -> WidgetState {
+        WidgetState {
             id,
             layout_rect: None,
             paint_insets: Insets::ZERO,
@@ -945,7 +945,7 @@ impl BaseState {
     /// Update to incorporate state changes from a child.
     ///
     /// This will also clear some requests in the child state.
-    fn merge_up(&mut self, child_state: &mut BaseState) {
+    fn merge_up(&mut self, child_state: &mut WidgetState) {
         let mut child_region = child_state.invalid.clone();
         child_region += child_state.layout_rect().origin().to_vec2() - child_state.viewport_offset;
         let clip = self
@@ -1017,10 +1017,10 @@ mod tests {
         let mut widget = WidgetPod::new(widget).boxed();
 
         let mut command_queue: CommandQueue = VecDeque::new();
-        let mut state = BaseState::new(WidgetId::next());
+        let mut state = WidgetState::new(WidgetId::next());
         let mut ctx = LifeCycleCtx {
             command_queue: &mut command_queue,
-            base_state: &mut state,
+            widget_state: &mut state,
             window_id: WindowId::next(),
             window: &WindowHandle::default(),
         };
@@ -1028,9 +1028,9 @@ mod tests {
         let env = Env::default();
 
         widget.lifecycle(&mut ctx, &LifeCycle::WidgetAdded, &None, &env);
-        assert!(ctx.base_state.children.may_contain(&ID_1));
-        assert!(ctx.base_state.children.may_contain(&ID_2));
-        assert!(ctx.base_state.children.may_contain(&ID_3));
-        assert_eq!(ctx.base_state.children.entry_count(), 7);
+        assert!(ctx.widget_state.children.may_contain(&ID_1));
+        assert!(ctx.widget_state.children.may_contain(&ID_2));
+        assert!(ctx.widget_state.children.may_contain(&ID_3));
+        assert_eq!(ctx.widget_state.children.entry_count(), 7);
     }
 }